qcacld-3.0: Fix to allocate memory for roam reason
Currently, No memeory is allocated for roam_reason in skb
which can lead to invalid memory access.

Fix is to allocate memory for roam reason in skb to avoid
invalid memory access.

CRs-Fixed: 2449019

+++ core/sme/inc/csr_api.h
@@ -1594,7 +1594,7 @@ typedef struct tagCsrRoamInfo {
 	int rx_rate;
 	tSirMacCapabilityInfo capability_info;
 	uint32_t rx_mc_bc_cnt;
-	uint8_t roam_reason;
+	uint16_t roam_reason;
 } tCsrRoamInfo;
 
 typedef struct tagCsrFreqScanInfo {
